refactor: move footer to generic components
