+++ apps/frontend/src/components/Generic/Footer/FooterBottom.tsx
