+++ apps/frontend/src/components/Generic/Footer/FooterDecoration.tsx
