+++ apps/frontend/src/components/Generic/Footer/constants.ts
