+++ apps/frontend/src/components/Generic/Footer/index.tsx
